Remove unneeded dependencies from configure.ac & xcomposite.pc
We only need Xfixes headers for the XserverRegion definition, don't
need to link against it directly or require clients link against it,
so only use CFLAGS from it in the Makefile, and only list it in
Requires.private in the .pc file.

Don't need libXext at all - no headers or functions from it are used.

Also clean up some extra appearances of *_CFLAGS & *_LIBS that aren't needed.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Reviewed-by: Dan Nicholson <dbn.lists@gmail.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>